# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "server",
    "description": "",
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/artists/{id}": {
      "get": {
        "tags": [
          "artists"
        ],
        "operationId": "get_artist",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Artist ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Artist detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "artists"
        ],
        "operationId": "update_artist",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Artist ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArtistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated artist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "delete": {
        "tags": [
          "artists"
        ],
        "operationId": "delete_artist",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Artist ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/artists/{id}/images": {
      "post": {
        "tags": [
          "artists"
        ],
        "operationId": "upload_artist_image",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Artist ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ImageUpload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Image uploaded and linked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistImageInfo"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Artist not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/artists/{id}/images/{asset_id}": {
      "delete": {
        "tags": [
          "artists"
        ],
        "operationId": "delete_artist_image",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Artist ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "asset_id",
            "in": "path",
            "description": "Asset ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "patch": {
        "tags": [
          "artists"
        ],
        "operationId": "update_artist_image_kind",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Artist ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "asset_id",
            "in": "path",
            "description": "Asset ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArtistImageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Kind updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistImageInfo"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/api/songs/{id}/images/{asset_id}": {
      "delete": {
        "tags": [
          "songs"
        ],
        "operationId": "delete_song_image",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Song ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "asset_id",
            "in": "path",
            "description": "Asset ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "patch": {
        "tags": [
          "songs"
        ],
        "operationId": "update_song_image_kind",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Song ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "asset_id",
            "in": "path",
            "description": "Asset ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSongImageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Kind updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongImageInfo"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtistImageInfo": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AssetInfo"
          },
          {
            "type": "object",
            "required": [
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "description": "Serialized [`ArtistImageKind`] value indicating the image's semantic role."
              }
            }
          }
        ],
        "description": "An image asset attached to an artist with its semantic role."
      },
...
      "ArtistLinkInfo": {
        "type": "object",
        "description": "An external link associated with an artist.",
        "required": [
          "id",
          "url",
          "kind"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier of the link record."
          },
          "kind": {
            "type": "string",
            "description": "Serialized [`ArtistLinkKind`] value indicating the platform."
          },
          "label": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional display override label."
          },
          "url": {
            "type": "string",
            "description": "Fully qualified URL."
          }
        }
      },
      "ArtistLinkInput": {
        "type": "object",
        "description": "Input for creating an external link on an artist.",
        "required": [
          "url",
          "kind"
        ],
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/ArtistLinkKind",
            "description": "Platform this link points to."
          },
          "label": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional display override label."
          },
          "url": {
            "type": "string",
            "description": "Fully qualified URL."
          }
        }
      },
      "ArtistLinkKind": {
        "type": "string",
        "description": "Valid platform kinds for an artist external link.",
        "enum": [
          "youtube",
          "website",
          "x",
          "instagram",
          "twitch",
          "other"
        ]
      },
      "ArtistResponse": {
        "type": "object",
        "description": "A full artist record returned by detail and mutation endpoints.",
        "required": [
          "id",
          "name",
          "images",
          "links",
          "song_count",
          "performance_count"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional freeform bio or description."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier."
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArtistImageInfo"
            },
            "description": "Images attached to this artist."
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArtistLinkInfo"
            },
            "description": "External links for this artist."
          },
          "name": {
            "type": "string",
            "description": "Display name."
          },
          "performance_count": {
            "type": "integer",
            "format": "int64",
            "description": "Number of performances in which this artist appears as a singer.",
            "minimum": 0
          },
          "song_count": {
            "type": "integer",
            "format": "int64",
            "description": "Number of songs for which this artist is credited as an original artist.",
            "minimum": 0
          }
        }
      },
...
      "AssetInfo": {
        "type": "object",
        "description": "Common fields describing a media asset, whether internally hosted or externally referenced.\n\nExactly one of `storage_url` or `external_url` is non-null per asset.",
        "required": [
          "asset_id"
        ],
        "properties": {
          "asset_id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier of the asset record."
          },
          "credits": {
            "type": [
              "string",
              "null"
            ],
            "description": "Attribution text for the content creator."
          },
          "external_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "Referenced external URL. Present for externally linked content only."
          },
          "source_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "Link to the original source post or stream."
          },
          "storage_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "CDN or file server URL. Present for internally hosted files only."
          },
          "title": {
            "type": [
              "string",
              "null"
            ],
            "description": "Display name or original filename. May be absent for unnamed uploads."
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Body returned for all error responses.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human readable description of the error."
          }
        }
      },
...
      "ImageUpload": {
        "type": "object",
        "description": "Placeholder schema for image multipart upload bodies.",
        "required": [
          "file",
          "kind"
        ],
        "properties": {
          "credits": {
            "type": [
              "string",
              "null"
            ]
          },
          "file": {
            "type": "string",
            "format": "binary"
          },
          "kind": {
            "type": "string",
            "description": "Semantic role of the image. See [`ArtistImageKind`](api_types::artists::ArtistImageKind)."
          },
          "source_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "SongImageInfo": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AssetInfo"
          },
          {
            "type": "object",
            "required": [
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "description": "Serialized [`SongImageKind`] value indicating the image's semantic role."
              }
            }
          }
        ],
        "description": "An image asset attached to a song with its semantic role."
      },
...
      "UpdateArtistImageRequest": {
        "type": "object",
        "description": "Request body for `PATCH /api/artists/{id}/images/{image_id}`.",
        "required": [
          "kind"
        ],
        "properties": {
          "kind": {
            "type": "string",
            "description": "Serialized [`ArtistImageKind`] value to assign."
          }
        }
      },
      "UpdateArtistRequest": {
        "type": "object",
        "description": "Request body for `PUT /api/artists/{id}`.\n\nImages are managed via the `/images` subresource.",
        "required": [
          "name",
          "links"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional freeform bio or description."
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArtistLinkInput"
            },
            "description": "Replaces all existing external links."
          },
          "name": {
            "type": "string",
            "description": "Display name."
          }
        }
      },
...
      "UpdateSongImageRequest": {
        "type": "object",
        "description": "Request body for `PATCH /api/songs/{id}/images/{image_id}`.",
        "required": [
          "kind"
        ],
        "properties": {
          "kind": {
            "type": "string",
            "description": "Serialized [`SongImageKind`] value to assign."
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    }
  }
}